refactor(stella-cli): split the loop's query verbs out of self_driving_cmd.rs
self_driving_cmd.rs sat at 1499 of its 1500-line ceiling, ungrandfathered
— one line of headroom, and the next change to touch it would have been
somebody else's, about something else.

The seam is the one the file had already drawn for itself. `aperture`,
`seen`, `calibrate_cmd` and `queue` sat under a banner comment reading
`aperture / seen / calibrate / queue`; this change only makes that
grouping a module boundary. None of the four drives a turn or touches a
pull request — each reports a facet of the loop's own state or nudges it
by hand, which is why they read together.

self_driving_cmd.rs 1499 -> 1388, with 112 lines of headroom. No
baseline entry: the file is not grandfathered and must not become so.

Second of the three files #5225 names; `drive.rs` was the first, and
`stella-tui/src/model.rs` is untouched.

Refs #5225
